ssh: rewrite ssh responder to use cache_req
This is a bigger change since both supported commands could be
rewritten for cache_req and the logic could be deleted. I decided
to also split the file into more modules and follow similar pattern
as with nss responder.

Resolves:
https://fedorahosted.org/sssd/ticket/1126
